--- packages/expo-asset/src/AssetRegistry.native.js
@@ -1,2 +1,0 @@
-import AssetRegistry from 'react-native/Libraries/Image/AssetRegistry';
-export default AssetRegistry;
