--- packages/expo-asset/src/AssetSourceResolver.native.js
@@ -1,2 +1,0 @@
-import AssetSourceResolver from 'react-native/Libraries/Image/AssetSourceResolver';
-export default AssetSourceResolver;
