--- packages/expo-asset/src/resolveAssetSource.native.js
@@ -1,2 +1,0 @@
-import resolveAssetSource from 'react-native/Libraries/Image/resolveAssetSource';
-export default resolveAssetSource;
